pyproject.toml
[project]
name = "medstudent"
version = "0.1.0"
description = ""
authors = [
    {name = "mcunha12",email = "marceloopcunha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "streamlit-authenticator (>=0.4.2,<0.5.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "streamlit-shadcn-ui (>=0.1.18,<0.2.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
packages = [{include = "medstudent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
